// src/api/schemas/pollListDetailsSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { PollListCommunitySchema } from './pollListCommunitySchema';
import type { PollListDetailsSchemaMaxPolls } from './pollListDetailsSchemaMaxPolls';
import type { PollListDetailsSchemaParentId } from './pollListDetailsSchemaParentId';
import type { PollListProfileSchema } from './pollListProfileSchema';

export interface PollListDetailsSchema {
  id: number;
  unique_id: string;
  slug: string;
  title: string;
  description: string;
  list_type: string;
  visibility: string;
  is_collaborative: boolean;
  is_featured: boolean;
  max_polls: PollListDetailsSchemaMaxPolls;
  direct_polls_count: number;
  total_polls_count: number;
  direct_folders_count: number;
  total_items_count: number;
  parent_id: PollListDetailsSchemaParentId;
  depth: number;
  order: number;
  path: string;
  profile: PollListProfileSchema;
  community: PollListCommunitySchema;
  can_edit?: boolean;
  can_add_polls?: boolean;
  is_owner?: boolean;
  view_count: number;
  like_count: number;
  bookmark_count: number;
  created_at: string;
  updated_at: string;
}